# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/signup/owner": {
      "post": {
        "description": "Crée le compte Supabase Auth et le profil propriétaire, envoie l'email de confirmation WARAH.",
        "operationId": "AuthController_signupOwner",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupOwnerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Inscription propriétaire",
        "tags": ["Auth"]
      }
    },
    "/auth/signup/manager": {
      "post": {
        "description": "Crée le compte Supabase Auth et le profil gestionnaire, envoie l'email de confirmation WARAH.",
        "operationId": "AuthController_signupManager",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupManagerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Inscription gestionnaire",
        "tags": ["Auth"]
      }
    },
    "/auth/invite/tenant": {
      "post": {
        "description": "Réservé au propriétaire du bien ou au gestionnaire mandaté (canActOnProperty). Crée immédiatement le compte locataire (email confirmé d'office) et envoie le lien d'activation par email — le locataire n'aura plus qu'à poser un mot de passe.",
        "operationId": "AuthController_inviteTenant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteTenantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Invite un locataire sur un bien",
        "tags": ["Auth"]
      }
    },
    "/auth/signup/tenant": {
      "post": {
        "description": "Vérifie le token signé reçu par email (paramètre `token`, expire après 7 jours) et pose le mot de passe du compte déjà créé — aucun autre champ à fournir.",
        "operationId": "AuthController_signupTenant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetTenantPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Active le compte locataire créé par l'invitation",
        "tags": ["Auth"]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Route la connexion via NestJS (plutôt que directement Supabase côté client) pour appliquer le blocage de 15 minutes après 5 tentatives échouées. Renvoie une session Supabase valide en cas de succès.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Connexion par email et mot de passe",
        "tags": ["Auth"]
      }
    },
...
    "/auth/password-reset/request": {
      "post": {
        "description": "Envoie un code à 6 chiffres par email (expire après 10 minutes, usage unique). Toute nouvelle demande invalide les codes précédents. Réponse toujours générique, qu'un compte existe ou non pour cet email.",
        "operationId": "AuthController_requestPasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Demande un code de réinitialisation de mot de passe",
        "tags": ["Auth"]
      }
    },
    "/auth/password-reset/confirm": {
      "post": {
        "operationId": "AuthController_confirmPasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmPasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Confirme la réinitialisation avec le code reçu par email",
        "tags": ["Auth"]
      }
    },
    "/auth/me": {
      "get": {
        "description": "Renvoie l'utilisateur courant avec son profil de rôle (Owner/Tenant/Manager/Admin), son statut de compte et ses préférences de notification.",
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Profil de l'utilisateur courant",
        "tags": ["Auth"]
      }
    },
...
    "/admin/stats": {
      "get": {
        "operationId": "AdminController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Statistiques globales plateforme (super admin)",
        "tags": ["Admin"]
      }
    },
    "/admin/users": {
      "get": {
        "operationId": "AdminController_listUsers",
        "parameters": [
          {
            "name": "role",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Liste tous les comptes (super admin)",
        "tags": ["Admin"]
      }
    },
    "/admin/users/{id}": {
      "get": {
        "operationId": "AdminController_getUserDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Détail d'un compte (super admin)",
        "tags": ["Admin"]
      },
      "delete": {
        "operationId": "AdminController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer (anonymiser) un compte (super admin)",
        "tags": ["Admin"]
      }
    },
    "/admin/reviews/{id}/moderate": {
      "patch": {
        "operationId": "AdminController_moderateReview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModerateReviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Masque ou réaffiche un avis signalé (super admin)",
        "tags": ["Admin"]
      }
    },
...
  },
  "info": {
    "title": "WARAH API",
    "description": "API de gestion locative WARAH — marché togolais",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SignupOwnerDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "example": "90330557"
          },
          "city": {
            "type": "string",
            "example": "Lomé"
          },
          "residenceCountry": {
            "type": "string",
            "example": "TG"
          }
        },
        "required": [
          "email",
          "password",
          "firstName",
          "lastName",
          "phone",
          "city",
          "residenceCountry"
        ]
      },
      "SignupManagerDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "example": "90330557"
          },
          "city": {
            "type": "string",
            "example": "Lomé"
          }
        },
        "required": ["email", "password", "firstName", "lastName", "phone", "city"]
      },
      "InviteTenantDto": {
        "type": "object",
        "properties": {
          "propertyId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "example": "90330557"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "monthlyRent": {
            "type": "number",
            "description": "Absent = reprend le loyer actuel du bien"
          },
          "monthlyCharges": {
            "type": "number",
            "description": "Absent = reprend les charges actuelles du bien"
          },
          "paymentFrequency": {
            "type": "string",
            "enum": ["MONTHLY", "QUARTERLY", "BIANNUAL", "ANNUAL"]
          },
          "startDate": {
            "type": "string",
            "description": "Date de début du bail (ISO 8601)"
          },
          "endDate": {
            "type": "string",
            "description": "Absent = bail ouvert, calendrier glissant sur 12 mois"
          },
          "securityDeposit": {
            "type": "number"
          },
          "depositReturnConditions": {
            "type": "string"
          },
          "reminderDaysBefore": {
            "type": "number",
            "description": "Jours avant échéance pour le rappel — omis = pas de rappel configuré"
          },
          "overdueAlertWindowDays": {
            "type": "number",
            "description": "Fenêtre (en jours) de l'alerte de retard après échéance — omis = pas d'alerte configurée"
          }
        },
        "required": [
          "propertyId",
          "email",
          "phone",
          "firstName",
          "lastName",
          "paymentFrequency",
          "startDate",
          "securityDeposit"
        ]
      },
      "SetTenantPasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          }
        },
        "required": ["password"]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": ["email", "password"]
      },
      "RequestPasswordResetDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": ["email"]
      },
      "ConfirmPasswordResetDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "example": "123456"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": ["email", "code", "newPassword"]
      },
...
      "ModerateReviewDto": {
        "type": "object",
        "properties": {
          "isHidden": {
            "type": "boolean",
            "description": "true = masque l’avis de l’annuaire public, false = le réaffiche"
          }
        },
        "required": ["isHidden"]
      },
...
    }
  }
}